Organize parameters.
